+++ src/ru/neosvet/lesson1/interfaces/Exercise.java
@@ -1,4 +1,4 @@
-package ru.neosvet.lesson1;
+package ru.neosvet.lesson1.interfaces;
 
 public interface Exercise {
     void train(Entity entity);
