+++ src/ru/neosvet/lesson1/interfaces/Runner.java
@@ -1,4 +1,4 @@
-package ru.neosvet.lesson1;
+package ru.neosvet.lesson1.interfaces;
 
 public interface Runner {
